requirements.txt
cryptography>=41.0.0
pycryptodome==3.19.1
colorama>=0.4.6
Pillow
tqdm
opencv-python 